Refactor PaymentController
Make `PaymentController` an interface and `StripePaymentController`
its implementation. This makes the implementation closed.
