+++ stripe/src/main/java/com/stripe/android/Stripe.kt
@@ -93,7 +93,7 @@ class Stripe internal constructor(
     ) : this(
         stripeRepository,
         stripeNetworkUtils,
-        PaymentController.create(context.applicationContext, stripeRepository, enableLogging),
+        StripePaymentController.create(context.applicationContext, stripeRepository, enableLogging),
         publishableKey,
         stripeAccountId
     )
